refactor(core): decouple KnowledgeVaultRepository from SQLite (Phase 1)
